Add files via upload
This MATLAB file uses a combination of high-pass filtering, wavelet denoising, and outlier removal based on signal quality index and dynamic time warping to pre-process SCG signals. The goal of this pre-processing is to prepare the signals for use with deep learning algorithms, which can be used to predict peak systolic velocity or diagnose aortic valve pathologies. The file also creates a stratified cross-validation dataset to be used with these algorithms.